Relax composition detection
Previously, a module was only considered to be composed if it was used
in a member expression containing "propTypes". E.g.

var Foo = require('Foo');
var A = React.createClass({
  propTypes: {
    ...Foo.propTypes
  }
});

but not

var Foo = require('Foo');
var A = React.createClass({
  propTypes: {
    ...Foo
  }
});

The latter is now allowed as well, because react-docgen doesn't actually
care about the characteristics of the module (see #9).

This also fixes a bug where `propTypes: Foo.propTypes` wasn't treated
correctly.