Add support for destructuring resolution.
Until now, the propTypeHandler would not be able to recognize that

```
var {PropTypes} = require('react');
```

refers to the React module, and thus not recognize the types properly.

This extends `resolveToValue` to return a MemberExpression if the path
resolves to an ObjectPattern. I.e. `bar` in

```
var {foo: {bar}} = baz;
bar;
```

would resolve to a path representing `baz.foo.bar`.