Make it easier to add data to Documentation
This changes the API of the Documentation object. `getDescription` and
`setDescription` are gone. It was replaced by `get(key)` and `set(key,
value)`. This allows custom handlers to add arbitrary data.

This also removes `props` from the result object if there aren't any.
This makes it easier to test for their existence (e.g. `if (doc.props) {...}`).