Add --resolver option (closes #42)
`--resolver` either accepts the name of a built-in resolver
(findExportedComponentDefinition, findAllComponentDefinitions) or a path
to a module that exports a resolver function.

With this option, react-docgen can be more easily customized. It is not
necessary anymore to create a custom script and use the API just to use
a different / custom resolver.