Add resolver "findAllExportedComponentDefinitions"
This resolver is basically a copy of the
"findExportedComponentDefinition" resolver inclusive the tests.
The most significant difference is, that this resolver does not throw
when it encounters more than one component definition.
It adds all found component definitions into an array (deduplicated)
and returns that array at the end of execution.