Find nested MemberExpressions / resolve commonJS properties
When an object doesn't have a "local identifier", for example because
it is directly assigned to "module.exports", we still want to be able
to find the member that we are looking for (i.e. "propTypes").

Before:
When directly assigning a React Component to "(module.)exports", the
correct members (i.e. "propTypes") could not be found.
```javascript
module.exports = function() { return (<i />); };
module.exports.propTypes = { a: PropTypes.string };
```
or
```javascript
module.exports.Foo = function() { return (<i />); };
module.exports.Foo.propTypes = { b: PropTypes.number };
```

After applying this commit, the "getMemberExpressionValuePath" utility
method will be able to find members defined on the same object as the
component definition that has been passed in or tries to find the member
defined on "(module.)exports" but does not check if the passed in
component definition is also being exported / assigned to module.exports