Try to resolve MemberExpression to initialization value
This change enables `resolveToValue` to handle `MemberExpression`s.
When it encounters a MemberExpression it tries to find the definition
of the root object of that expression and then tries to resolve the
property inside that ObjectExpression.
If it cannot resolve the root object to an ObjectExpression
(ie. because it was declared externally), it will stop and return the
path to the original MemberExpression - or if the original expression
did resolve to an Object but the properties inside that object were
Identifier which could not resolve to an ObjectExpression it returns
that MemberExpression instead.