Make resolveToValue consider assignments to identifiers (fixes #58)
In the following example, resolveToValue wouldn't resolve to the value
assigned to "React" because it only looks at the variable
definition, not at assignments to the variable:

```
var React;
React = require('react');
module.exports = React.createClass(...);
```

This diff fixes that. It will make a shallow traversal over the scope
where the variable is defined and look for the last assignment to the
variable.

---

I'm also trying to improve tests by storing component definitions that
are known to have been broken in `src/__tests__/fixtures/`. Adding more
components over time will strengthen the test suit and help catch
regressions.