[RFC] Speed up react-docgen by 25-40%
I did some profiling and noticed that recast has a significant amount of overhead due to the fact that it copies every tree it parses and has complicated printing logic. I wasn't sure if react-docgen uses any mutative recast features so I tried to remove it.

The good news: this is about 25-40% faster in my testing (40% faster when run on 20 files, 25% faster when run on 700 components from FB internal repos -- probably the JIT mitigates some of the slowness of recast).

The bad news: this doesn't actaully work, because (surprise, surprise) we actually do use some of the modification features of recast. When these are used, we don't know how to reprint.

```
$ ag --ignore __tests__ builders\\.
src/utils/normalizeClassDefinition.js
79:            var classProperty = builders.classProperty(

src/utils/resolveObjectKeysToArray.js
98:        .map(value => builders.literal(value))
100:      return new NodePath(builders.arrayExpression(nodes));

src/utils/resolveToValue.js
33:        builders.memberExpression(
```

Maybe we could reimplement a small subset of the printing logic? Or maybe that's too impractical. FWIW all but one of the 700 files I tested produced the exact same output including whitespace in the `raw` JS source fields.